+++ src/test/java/ampControl/model/training/data/iterators/preprocs/CnnToManyToOneRnnPreProcessorTest.java
@@ -6,7 +6,8 @@ import org.nd4j.linalg.dataset.api.DataSet;
 import org.nd4j.linalg.dataset.api.DataSetPreProcessor;
 import org.nd4j.linalg.factory.Nd4j;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Test cases for {@link CnnToManyToOneRnnPreProcessor}.
@@ -20,10 +21,10 @@ public class CnnToManyToOneRnnPreProcessorTest {
      */
     @Test
     public void preProcess() {
-        final int[] featureShapeIn = {3,1,7,11};
-        final int[] expectedFeatureShape = {3,11,7};
-        final int[] labelsShapeIn = {3,5};
-        final int[] expectedLabelsShape = {3,5,7};
+        final long[] featureShapeIn = {3,1,7,11};
+        final long[] expectedFeatureShape = {3,11,7};
+        final long[] labelsShapeIn = {3,5};
+        final long[] expectedLabelsShape = {3,5,7};
         final double[] expectedMask = {0,0,0,0,0,0,1};
         final INDArray expectedLabelsMask = Nd4j.create(new double[][] {expectedMask, expectedMask, expectedMask});
         final DataSet testSet = new org.nd4j.linalg.dataset.DataSet(Nd4j.create(featureShapeIn), Nd4j.create(labelsShapeIn));
@@ -40,8 +41,8 @@ public class CnnToManyToOneRnnPreProcessorTest {
      */
     @Test
     public void cnnToRnnFeature() {
-        final int[] featureShapeIn = {3,1,7,11};
-        final int[] expectedFeatureShape = {3,11,7};
+        final long[] featureShapeIn = {3,1,7,11};
+        final long[] expectedFeatureShape = {3,11,7};
         assertArrayEquals("Incorrect shape!", expectedFeatureShape,
                 CnnToManyToOneRnnPreProcessor.cnnToRnnFeature(Nd4j.create(featureShapeIn)).shape());
     }
