+++ src/test/java/ampControl/model/training/model/layerblocks/graph/DummyOutputLayer.java
@@ -2,7 +2,6 @@ package ampControl.model.training.model.layerblocks.graph;
 
 import ampControl.model.training.model.layerblocks.LayerBlockConfig;
 import ampControl.model.training.model.layerblocks.adapters.BuilderAdapter;
-import org.deeplearning4j.nn.conf.distribution.BinomialDistribution;
 import org.deeplearning4j.nn.conf.layers.OutputLayer;
 import org.deeplearning4j.nn.graph.ComputationGraph;
 import org.deeplearning4j.nn.weights.WeightInit;
@@ -37,7 +36,7 @@ public class DummyOutputLayer implements LayerBlockConfig {
      * @param graph
      */
     public static void setEyeOutput(ComputationGraph graph) {
-        final int[] wShape = graph.getOutputLayer(0).getParam("W").shape();
+        final long[] wShape = graph.getOutputLayer(0).getParam("W").shape();
         graph.getOutputLayer(0).setParam("W", Nd4j.eye(wShape[0]));
     }
 }
